+++ src/index.ts
@@ -1,50 +1,14 @@
-import {
-  getCredentials,
-  getEvents,
-  initializeOAuth2Client,
-} from './util/googleApiUtility';
-import {
-  displayDateTimeRange,
-  groupEventsByDate,
-} from './util/dateTimeUtility';
+import { huntCommandHandler } from './hunt';
 import { fixCommandHandler } from './fix';
 import { exit } from 'process';
 import { displayHowToUse } from './help';
 import { initializeI18n } from './translation';
 
-const listEvents = async (eventName: string) => {
-  const oauth2Client = await initializeOAuth2Client();
-
-  try {
-    const events = await getEvents(oauth2Client, eventName);
-    if (events) {
-      const groupedEvents = groupEventsByDate(events);
-      if (groupedEvents.length > 0) {
-        console.log('Upcoming events:');
-        await displayDateTimeRange(groupedEvents);
-        if (events.length > 10) {
-          console.log('The number of events exceeds 10.');
-        }
-      } else {
-        console.log('No upcoming events found.');
-      }
-    } else {
-      console.log('Could not found event.');
-    }
-  } catch (error) {
-    console.log(error);
-    await getCredentials(oauth2Client);
-    await listEvents(eventName);
-  } finally {
-    exit();
-  }
-};
-
 initializeI18n();
 
 switch (process.argv[2]) {
   case 'hunt':
-    listEvents(process.argv[3]);
+    huntCommandHandler(process.argv[3]);
     break;
   case 'fix':
     fixCommandHandler(process.argv[3], process.argv[4], process.argv[5]);
